requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
fastapi==0.110.1
gunicorn==22.0.0
idna==3.7
mysql-connector-python==8.3.0
packaging==24.0
pydantic==2.7.0
pydantic-core==2.18.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
